# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YEAH API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/event": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "ambil data event",
                "description": "ambil data event",
                "operationId": "event",
                "responses": {
                    "200": {
                        "description": "sukses ambil data event"
                    },
                    "400": {
                        "description": "gagal mendapatkan data event"
                    }
                }
            }
        },
        "/api/v1/event/{id_events}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "ambil data event detail by id",
                "description": "ambil data event detail by id",
                "operationId": "eventdetail",
                "parameters": [
                    {
                        "name": "id_events",
                        "in": "path",
                        "description": "input id event",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sukses ambil data event detail by id"
                    },
                    "400": {
                        "description": "gagal mendapatkan data event detail by id"
                    }
                }
            }
        },
        "/api/v1/event/cekticket": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "ambil data ticket by booking code",
                "description": "ambil data ticket by booking code",
                "operationId": "cekticket",
                "parameters": [
                    {
                        "name": "code_booking",
                        "in": "path",
                        "description": "input booking code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "sukses ambil data ticket by booking code"
                    },
                    "400": {
                        "description": "gagal mendapatkan data ticket by booking code"
                    }
                }
            }
        }
    }
}